fix: 댓글 번역 커스텀 훅 추가
